src/crawl/requirements.txt
scrapy
requests
colorlog
confluent-kafka
scrapy-redis
python-dotenv